+++ spec/mailers/feedback_messages/further_information_request_feeback_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe FeedbackMessages::FurtherInformationRequestFeedback do
+RSpec.describe Nsm::FeedbackMessages::FurtherInformationRequestFeedback do
   subject(:feedback) { described_class.new(claim, caseworker_information_requested) }
 
   let(:claim) { build(:claim) }
